Feat:
Feat:
#6439
If the operator is in the lead and is placed in training, he remains in the lead.

When opting for a new level, the operator remains leader.

#6440

Validated by assigning level I and the user is already in training, Level I is assigned and the training level is updated to L Training.